Change getWinnings return type from int to double

+++ lab4/Raceable.java
@@ -2,5 +2,5 @@ package lab4;
 
 public interface Raceable {
     double averageWinnings();
-    int getWinnings();
+    double getWinnings();
 }
